+++ src/api/app.js
@@ -1,0 +1,84 @@
+import cors from 'cors'
+import express from 'express'
+import mongoose from 'mongoose'
+import {  MONGO_URI, SERVER_URL, PORT } from '../const.js'
+import notes from './notes/notes.js'
+
+export let connected = {}
+const app = express()
+app.use(cors())
+app.use(express.json())
+app.use('/', notes)
+
+app.listen(PORT, () => {
+  console.log(`Server running at ${SERVER_URL}`)
+})
+
+/**
+ * Create a new mongodb collection
+ */
+app.get(SERVER_URL + '/create', async (req, res) => {
+  const { dbName } = req.body
+  try {
+    const result = await create(dbName)
+    res.status(200).send(result)
+  } catch (err) {
+    res.status(500).send(err.message)
+  }
+})
+
+/**
+ * Connect to a dbName collection into the database "selfie"
+ *
+ * @param {*} dbName collection name to connect to
+ */
+export async function connect(dbName) {
+  try {
+    await mongoose.connect(MONGO_URI, { useNewUrlParser: true, useUnifiedTopology: true })
+    console.log('MongoDB Connected to ' + dbName + ' ...')
+    connected[dbName] = true
+  } catch (err) {
+    console.log(err)
+    connected[dbName] = false
+  }
+}
+
+/**
+ * Searching documents from a collection
+ *
+ * @param dbName collection name
+ */
+async function search(dbName) {
+  console.log('Getting docs from ' + dbName + ' ...')
+  if (!connected[dbName]) await connect(dbName)
+  else mongoose.create(dbName)
+  return mongoose.model(dbName).find()
+}
+
+/**
+ * Removing documents from a collection by ID
+ */
+app.post('/delete', async (req, res) => {
+  const { collection, id } = req.body
+  try {
+    await mongoose.model(collection).findByIdAndDelete(id)
+    res.status(200).send('Document deleted from ' + collection)
+  } catch (err) {
+    res.status(500).json({ error: err.message })
+  }
+})
+
+/**
+ * Disconnect from MongoDB
+ *
+ * @param {*} dbName
+ */
+async function disconnect(dbName) {
+  try {
+    await mongoose.disconnect()
+    console.log('MongoDB Disconnected from ' + dbName + ' ...')
+    connected[dbName] = false
+  } catch (err) {
+    console.log(err)
+  }
+}
