+++ src/api/notes/notes.js
@@ -1,29 +1,8 @@
-import cors from 'cors'
 import express from 'express'
 import mongoose from 'mongoose'
-import { MONGO_URI, SERVER_URL, PORT } from '../../src/const.js'
+import { connected, connect } from '../app.js'
 
-let connected = {}
 const app = express()
-app.use(cors())
-app.use(express.json())
-
-app.listen(PORT, () => {
-  console.log(`Server running at ${SERVER_URL}`)
-})
-
-/**
- * Create a new mongodb collection
- */
-app.get(SERVER_URL + '/create', async (req, res) => {
-  const { dbName } = req.body
-  try {
-    const result = await create(dbName)
-    res.status(200).send(result)
-  } catch (err) {
-    res.status(500).send(err.message)
-  }
-})
 
 /**
  * From 'notes' collection get all documents
@@ -56,21 +35,6 @@ app.post('/search', async (req, res) => {
   }
 })
 
-/**
- * Connect to a dbName collection into the database "selfie"
- *
- * @param {*} dbName collection name to connect to
- */
-async function connect(dbName) {
-  try {
-    await mongoose.connect(MONGO_URI, { useNewUrlParser: true, useUnifiedTopology: true })
-    console.log('MongoDB Connected to ' + dbName + ' ...')
-    connected[dbName] = true
-  } catch (err) {
-    console.log(err)
-    connected[dbName] = false
-  }
-}
 
 const NoteSchema = new mongoose.Schema({
   name: {
@@ -135,7 +99,7 @@ app.post('/save', async (req, res) => {
 })
 
 // Ottieni i tag di una nota specifica
-app.get('/notes/:id/tags', async (req, res) => {
+app.get('/:id/tags', async (req, res) => {
   try {
     const note = await Note.findById(req.params.id)
     if (note == null) {
@@ -148,7 +112,7 @@ app.get('/notes/:id/tags', async (req, res) => {
 })
 
 // Aggiungi un tag a una nota esistente
-app.post('/notes/:id/tags', async (req, res) => {
+app.post('/:id/tags', async (req, res) => {
   try {
     const note = await Note.findById(req.params.id)
     if (note == null) {
@@ -176,51 +140,11 @@ app.get('/tags', async (req, res) => {
 })
 
 /**
- * Searching documents from a collection
- *
- * @param dbName collection name
- */
-async function search(dbName) {
-  console.log('Getting docs from ' + dbName + ' ...')
-  if (!connected[dbName]) await connect(dbName)
-  else mongoose.create(dbName)
-  return mongoose.model(dbName).find()
-}
-
-/**
- * Removing documents from a collection by ID
- */
-app.post('/delete', async (req, res) => {
-  const { collection, id } = req.body
-  try {
-    await mongoose.model(collection).findByIdAndDelete(id)
-    res.status(200).send('Document deleted from ' + collection)
-  } catch (err) {
-    res.status(500).json({ error: err.message })
-  }
-})
-
-/**
- * Disconnect from MongoDB
- *
- * @param {*} dbName
- */
-async function disconnect(dbName) {
-  try {
-    await mongoose.disconnect()
-    console.log('MongoDB Disconnected from ' + dbName + ' ...')
-    connected[dbName] = false
-  } catch (err) {
-    console.log(err)
-  }
-}
-
-/**
  *   ****** UPLOAD IMAGE ******   *
  */
 
 import bodyParser from 'body-parser'
-import upload from '../upload.js'
+import upload from './upload.js'
 
 const ImageSchema = new mongoose.Schema({
   filename: String,
@@ -255,3 +179,5 @@ app.post('/upload', upload.single('image'), async (req, res) => {
     res.status(500).json({ success: 0, message: 'Upload image error!' })
   }
 })
+
+export default app
